requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blessed==1.20.0
blobfile==2.1.1
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.1
comm==0.2.1
contourpy==1.1.1
cycler==0.12.1
DALL-E==0.1
datasets==2.18.0
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
docker-pycreds==0.4.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fonttools==4.48.1
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.27.0
google-auth-oauthlib==1.0.0
gpustat==1.1.1
grpcio==1.60.1
h5py==3.11.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.29.2
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jupyter-client==8.6.0
jupyter-core==5.7.1
kiwisolver==1.4.5
lightning==2.2.0.post0
lightning-utilities==0.10.1
lit==17.0.6
lxml==4.9.4
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.7.4
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy==1.8.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.133
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycryptodomex==3.20.0
pygments==2.17.2
pyparsing==3.1.1
pytest==8.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytorch-lightning==2.2.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
scipy==1.14.1
seaborn==0.13.2
sentry-sdk==1.45.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
timm==0.9.12
tokenizers==0.15.2
tomli==2.0.1
torch==2.0.1
torch-tb-profiler==0.4.3
torchaudio==2.0.2
torchmetrics==1.3.1
torchmultimodal-nightly==2024.2.13
torchvision==0.15.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.2
triton==2.0.0
typing-extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
wandb==0.16.6
wcwidth==0.2.13
werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
